+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Indexes can be separated in the same line by ',' and ';' as well as ' '. Empty entries will be removed.
+
 ### Changed
 
 - Output folder opening with a button instead of automatically after saving file.